Sever parsing code from rest of cros-codecs
Also mark almost all dependencies as optional and disable their
compilation if we just want the parser stack.
