+++ Cargo.toml
@@ -9,18 +9,18 @@ edition = "2021"
 
 [features]
 default = ["vaapi"]
-vaapi = ["libva"]
-v4l2 = ["v4l2r"]
+vaapi = ["libva", "anyhow", "byteorder", "thiserror", "crc32fast", "nix"]
+v4l2 = ["v4l2r", "anyhow", "byteorder", "thiserror", "crc32fast", "nix"]
 
 [dependencies]
-anyhow = "1"
-byteorder = "1.4.3"
+anyhow = { version = "1", optional = true }
+byteorder = { version = "1.4.3", optional = true }
 libva = { version = "0.0.7", package = "cros-libva", optional = true }
 v4l2r = { version = "0.0.3", package = "v4l2r", optional = true }
-log = { version = "0", features = ["release_max_level_debug"] }
-thiserror = "1.0.31"
-crc32fast = "1.3.2"
-nix = { version = "0.29", features = ["fs", "event"] }
+log = { version = "0", features = ["release_max_level_off"] }
+thiserror = { version = "1.0.31", optional = true }
+crc32fast = { version = "1.3.2", optional = true }
+nix = { version = "0.29", features = ["fs", "event"], optional = true }
 
 [dev-dependencies]
 argh = "0.1"
