+++ examples/ccdec.rs
@@ -36,9 +36,9 @@ use cros_codecs::utils::simple_playback_loop;
 use cros_codecs::utils::simple_playback_loop_owned_frames;
 use cros_codecs::utils::simple_playback_loop_userptr_frames;
 use cros_codecs::utils::DmabufFrame;
-use cros_codecs::utils::IvfIterator;
-use cros_codecs::utils::NalIterator;
 use cros_codecs::utils::UserPtrFrame;
+use cros_codecs::bitstream_utils::IvfIterator;
+use cros_codecs::bitstream_utils::NalIterator;
 use cros_codecs::DecodedFormat;
 use cros_codecs::Fourcc;
 use cros_codecs::FrameLayout;
