+++ examples/ccenc.rs
@@ -22,8 +22,8 @@ use cros_codecs::encoder::stateless::vp9;
 use cros_codecs::encoder::vp9::EncoderConfig as VP9EncoderConfig;
 use cros_codecs::encoder::FrameMetadata;
 use cros_codecs::encoder::VideoEncoder;
-use cros_codecs::utils::IvfFileHeader;
-use cros_codecs::utils::IvfFrameHeader;
+use cros_codecs::bitstream_utils::IvfFileHeader;
+use cros_codecs::bitstream_utils::IvfFrameHeader;
 use cros_codecs::BlockingMode;
 use cros_codecs::Fourcc;
 use cros_codecs::FrameLayout;
