+++ src/decoder/stateless.rs
@@ -28,7 +28,6 @@ use nix::sys::epoll::EpollFlags;
 use nix::sys::eventfd::EventFd;
 use thiserror::Error;
 
-use crate::codec::vp8::parser::ParseFrameError;
 use crate::decoder::BlockingMode;
 use crate::decoder::DecodedHandle;
 use crate::decoder::DecoderEvent;
