+++ src/decoder/stateless/vp8/vaapi.rs
@@ -321,7 +321,7 @@ mod tests {
     use crate::decoder::BlockingMode;
     use crate::utils::simple_playback_loop;
     use crate::utils::simple_playback_loop_owned_frames;
-    use crate::utils::IvfIterator;
+    use crate::bitstream_utils::IvfIterator;
     use crate::DecodedFormat;
     use crate::Resolution;
