+++ src/decoder/stateless/vp9/vaapi.rs
@@ -327,7 +327,7 @@ mod tests {
     use crate::decoder::BlockingMode;
     use crate::utils::simple_playback_loop;
     use crate::utils::simple_playback_loop_owned_frames;
-    use crate::utils::IvfIterator;
+    use crate::bitstream_utils::IvfIterator;
     use crate::DecodedFormat;
 
     use super::*;
