+++ src/encoder/stateless/av1/vaapi.rs
@@ -621,8 +621,8 @@ mod tests {
     use crate::encoder::FrameMetadata;
     use crate::encoder::RateControl;
     use crate::encoder::Tunings;
-    use crate::utils::IvfFileHeader;
-    use crate::utils::IvfFrameHeader;
+    use crate::bitstream_utils::IvfFileHeader;
+    use crate::bitstream_utils::IvfFrameHeader;
     use crate::FrameLayout;
     use crate::PlaneLayout;
     use crate::Resolution;
